+++ midix/src/file/meta/mod.rs
@@ -6,3 +6,5 @@ mod time_signature;
 pub use time_signature::*;
 mod key_signature;
 pub use key_signature::*;
+mod ref_type;
+pub use ref_type::*;
